Refactor `uru admin refresh` behavior. Fixes #39
Uru automagically supports two registry refresh styles. The default
is useful when the registered ruby is never updated once registered,
or the custom tag label should remain the same across updates. The
second style is useful when a registered ruby is regularly updated to
the latest patch level and the default generated tag label should be
updated to match. The two styles are invoked as:

  uru admin refresh           # default refresh style
  uru admin refresh --retag   # retag refresh style

Additional refresh and custom tag labeling is supported with the
manual use of the `uru admin retag` command.