Add initial chocolatey package build support
This commit is based upon revisions to Thermatix's initial code
submission. Thank you.

The primary task is `choco:package` which dynamically creates the
required *.nuspec and *.ps1 files in the `build/chocolatey` dir.
The task then builds the *.nupkg file and moves it to the `pkg` dir
after a successful build.