Use "tag hash" rather than overloading "tag"
Uru internally uses a hash of metadata to uniquely identify each
registered ruby. This "tag hash" is a non-user visible token created
when a user registers a particular ruby. The tag hash is different
than the user created and visible "tag label", and is different than
the internal "tag" concept.

As such, use better variable names and change the name of helper func

    PathListForTag(...) -> PathListForTagHash(...)