Fix weakness with PATH canaries. Closes #90, #92
Uru's use of relative PATH canaries `_U1_` and `_U2_` to sandbox
PATH to enable easy activation and deactivation of registered rubies
was a security weakness. This commit minimizes that weakness by
using absolute PATH canaries `/_U1_` and `/_U2_` in Windows, Linux,
and OSX environments. In MSYS2-based systems on Windows such as
git-bash, `U:\_U1_` and `U:\_U2_` are used.